requirements.txt
bs4
langchain
langchain-text-splitters
langchain-community
langgraph
langchain-groq
langchain-openai
langchain-core
langchain-nomic
python-dotenv
pypdf2
termcolor
tiktoken
langchain-postgres 
psycopg2-binary 
supabase
supabase-py
python-jose[cryptography]
postgrest
httpx
